+++ app/components/mandatarissen/history-row.hbs
@@ -42,7 +42,7 @@
     {{else}}
       {{#if @mandataris.id}}
         <AuLink
-          @route="mandatarissen.mandataris"
+          @route="mandatarissen.persoon.mandataris"
           @model={{@mandataris.id}}
           @skin="secondary"
         >
