+++ app/components/mandatarissen/mandataris-card.hbs
@@ -71,7 +71,7 @@
           <div class="au-c-description-value">
             {{#each @vervangers as |replaced index|}}
               <AuLink
-                @route="mandatarissen.mandataris"
+                @route="mandatarissen.persoon.mandataris"
                 @model={{replaced.id}}
                 @skin="secondary"
               >
@@ -92,7 +92,7 @@
           <div class="au-c-description-value">
             {{#each @mandataris.tijdelijkeVervangingen as |replacement index|}}
               <AuLink
-                @route="mandatarissen.mandataris"
+                @route="mandatarissen.persoon.mandataris"
                 @model={{replacement.id}}
                 @skin="secondary"
               >
