+++ app/components/mandatarissen/mandataris-summary.hbs
@@ -6,7 +6,7 @@
   </div>
   <div class="au-o-grid__item au-u-2-5 au-u-text-right">
     <AuLink
-      @route="mandatarissen.mandataris"
+      @route="mandatarissen.persoon.mandataris"
       @model={{@mandataris.id}}
       @skin="button"
     >
