+++ app/controllers/mandatarissen/persoon/index.js
@@ -3,7 +3,7 @@ import { action } from '@ember/object';
 import { tracked } from '@glimmer/tracking';
 import { inject as service } from '@ember/service';
 
-export default class MandatarissenPersoonController extends Controller {
+export default class MandatarissenPersoonIndexController extends Controller {
   @service router;
 
   @tracked isModalOpen = false;
