+++ app/templates/mandatarissen/persoon/mandataris.hbs
@@ -1,6 +1,6 @@
 {{breadcrumb
   this.model.mandataris.bekleedt.bestuursfunctie.label
-  route="mandatarissen.mandataris"
+  route="mandatarissen.persoon.mandataris"
   model=this.model.mandataris.id
 }}
