+++ app/templates/organen/orgaan/mandatarissen.hbs
@@ -35,7 +35,7 @@
   @page={{this.page}}
   @size={{this.size}}
   @personRoute="mandatarissen.persoon"
-  @mandaatRoute="mandatarissen.mandataris"
+  @mandaatRoute="mandatarissen..persoon.mandataris"
 />
 
 {{outlet}}
